+++ snaps/cacher/utils/__init__.py
@@ -1,8 +1,10 @@
 from .key_gen import generate_auto_key, generate_template_key
 from .policy_require_checker import check_policy_requirements
+from .protocols import SnapFunction
 
 __all__ = [
     "generate_auto_key",
     "generate_template_key",
     "check_policy_requirements",
+    "SnapFunction",
 ]